rename file name and remove plug from resp ring init
